+++ Packages/BrandonUtils/CHANGELOG.md
@@ -5,16 +5,31 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.4.0] - 6/29/2021
+
+### Added
+
+- `RegexUtils`
+- `CollectionUtils.None()`
+- `EnumUtils.Parse()`
+- `Optional.IfOrElse()`
+
 ## [2.3.1] - 6/29/2021
 
-- Added trimmed stack traces to `AssertAll.Of()`
-- Added words to the dictionary
+### Added
+
+- Trimmed stack traces to `AssertAll.Of()`
+- Words to the dictionary
 
 ## [2.3.0] - 6/29/2021
+
+### Added
 
 - `PrimitiveUtils` - bold extensions for "primitive" types like `int`
 
 ## [2.2.0] - 6/29/2021
+
+### Added
 
 - New `StringUtils`
     - `SplitLines()` - splits a string by line breaks